backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aiosqlite
pydantic
pydantic-settings
alembic
python-dotenv
greenlet
